backend/requirements.txt
asgiref==3.9.1
Django==5.2.6
django-filter==25.1
djangorestframework==3.16.1
Markdown==3.8.2
sqlparse==0.5.3
